Register correlationId before reserve request is made
Otherwise there is a race condition where the provider may already
have send the reply before the web browser connects to the event
source. In this case the reply would not have matched any existing
correlation id. By registering the correlation id first it will always
exist when the reply comes in or the browser subscribes to the event
source.